MathPotato::ast::ast_tree::i32_api_put_node

Trait I32ApiPutNode

pub trait I32ApiPutNode: I32Api {
    // Required method
    fn put_i32_ast_node(
        &mut self,
        node: I32AstNode,
    ) -> Result<I32AstNode, ParseError>;
}

Required Methods

fn put_i32_ast_node( &mut self, node: I32AstNode, ) -> Result<I32AstNode, ParseError>

Adds a new I32AstNode to the AST.

Remarks

By adding a new node to the AST the reference to the continuation node will be updated accordingly.

When there is a I32AstNode in the AST with the same key there will be no overwrite, rather error message.

Parameters

node - an I32AstNode

Returns

OK(I32AstNode) - the newly created I32AstNode node which also has the unique identifier.

Err(ParseError) - when recording the node is not successful due to some reasons.