MathPotato::ast::ast_tree::i32_api_update_node

Trait I32ApiUpdateNode

pub trait I32ApiUpdateNode {
    // Required method
    fn update_i32_node(
        &mut self,
        id: Uuid,
        node: I32AstNode,
    ) -> Result<I32AstNode, ParseError>;
}

The I32 type related functionalities offered by the Abstract Syntax Tree.

Required Methods

fn update_i32_node( &mut self, id: Uuid, node: I32AstNode, ) -> Result<I32AstNode, ParseError>

Updates the designated i32 datatype node with the provided data.

Details

If there is no i32 node with the provided id the operation will panic.

Parameters

id - Uuid: The unique id of the node to be updated.

node - I32AstNode: The I32AstNode with the new data.