MathPotato::ast::ast_tree::root_node_api

Trait RootNodeApi

pub trait RootNodeApi {
    // Required methods
    fn add_root_node_id_and_type(
        &mut self,
        id: Uuid,
        node_type: AstNodeType,
    ) -> Result<(Uuid, AstNodeType), ParseError>;
    fn get_root_node_infix(
        &self,
        id: Uuid,
    ) -> Result<(Uuid, InfixAstNodeInternal), ParseError>;
    fn update_root_node_id_and_type(
        &mut self,
        id: Uuid,
        node_type: AstNodeType,
    ) -> Result<(Uuid, AstNodeType), ParseError>;
    fn get_root_node_id(&self) -> Option<Uuid>;
    fn get_root_node_type(&self) -> AstNodeType;
}

Root Node Api of the Abstract Syntax Tree

The root node of the AST must be tracked for the interpreter which starts here and walks the tree and does its magic. At this point I think there will be a single root node.

Required Methods

fn add_root_node_id_and_type( &mut self, id: Uuid, node_type: AstNodeType, ) -> Result<(Uuid, AstNodeType), ParseError>

fn get_root_node_infix( &self, id: Uuid, ) -> Result<(Uuid, InfixAstNodeInternal), ParseError>

Returns the root node by id.

fn update_root_node_id_and_type( &mut self, id: Uuid, node_type: AstNodeType, ) -> Result<(Uuid, AstNodeType), ParseError>

fn get_root_node_id(&self) -> Option<Uuid>

fn get_root_node_type(&self) -> AstNodeType