MathPotato::ast::ast_tree

Trait I32

pub trait I32 {
    // Required methods
    fn get_i32_node_by_id(&self, id: Uuid) -> Result<I32AstNode, ParseError>;
    fn put_i32_ast_node(
        &mut self,
        node: I32AstNode,
    ) -> Result<I32AstNode, ParseError>;
    fn i32_tree_len(&self) -> usize;
    fn update_i32_node(
        &mut self,
        id: Uuid,
        node: I32AstNode,
    ) -> Result<I32AstNode, ParseError>;
}

The I32 type related functionalities offered by the Abstract Syntax Tree.

Required Methods

fn get_i32_node_by_id(&self, id: Uuid) -> Result<I32AstNode, ParseError>

Returns the I32AstNode with the designated id.

Parameters

id - Uuid - the id of the requested node

Returns

Result<I32AstNode, ParseError

fn put_i32_ast_node( &mut self, node: I32AstNode, ) -> Result<I32AstNode, ParseError>

Adds a new node to the AST. By adding a new node to the AST the reference to the continuation node will be updated accordingly.

Important

When there is a I32AstNode in the AST with the same key there will be no overwrite, rather error message.

Parameters

node - an I32AstNode

Returns

Result<(Uuid, I32AstNode), ParseError>

fn i32_tree_len(&self) -> usize

Returns the length of the i32 datatype tree size.

fn update_i32_node( &mut self, id: Uuid, node: I32AstNode, ) -> Result<I32AstNode, ParseError>

Updates the designated i32 datatype node with the provided data.

Details

If there is no i32 node with the provided id the operation will panic.

Parameters

id - Uuid: The unique id of the node to be updated.

node - I32AstNode: The I32AstNode with the new data.